+++ PodcastMagazine/Screens/SearchView/ResultContentView.swift
@@ -45,76 +45,3 @@ struct ResultContentView: View {
 #Preview {
     ResultContentView(userQuery: "Podlodka")
 }
-
-struct ResultView: View {
-    var query: String
-    let items: [Feed]
-    let podcastItems: [Feed]
-    
-    init(
-        query: String,
-        items: [Feed],
-        podcastItems: [Feed]
-    ) {
-        self.query = query
-        self.items = items
-        self.podcastItems = podcastItems
-    }
-    
-    private var rows: [GridItem] = [GridItem(.flexible())]
-    
-    var body: some View {
-        
-        VStack {
-            HeaderResultView(title: query)
-            
-            VStack {
-                ScrollView(showsIndicators: false) {
-                    LazyVGrid(columns: rows,
-                              alignment: .center,
-                              spacing: nil,
-                              pinnedViews: [],
-                              content: {
-                        Section {
-                            ForEach(items, id: \.id) { item in
-                                ResultCellView(
-                                    isResultCell: true, 
-                                    item: item
-                                )
-                                    .frame(height: 72)
-                            }
-                        } header: {
-                            HStack {
-                                Text("Search Result")
-                                    .foregroundStyle(Color.mainText)
-                                    .font(.custom(.extraBold, size: 14))
-                                Spacer()
-                            }
-                        }
-                        
-                        Section {
-                            ForEach(podcastItems, id: \.id) { item in
-                                ResultCellView(
-                                    isResultCell: false,
-                                    item: item
-                                )
-                                    .frame(height: 72)
-                            }
-                        } header: {
-                            HStack {
-                                Text("All Podcast")
-                                    .foregroundStyle(Color.mainText)
-                                    .font(.custom(.medium, size: 14))
-                                Spacer()
-                            }
-                        }
-                    })
-                    .padding(.top, 24)
-                }
-                .padding(.horizontal, 33)
-            }
-        }
-        .navigationBarHidden(true)
-        .padding(.top, 49)
-    }
-}
